forwardgram-master/requirements.txt
telethon
pyyaml
PySocks